requirements.txt
# Slum Detection Model Requirements
# =================================

# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Segmentation Models
segmentation-models-pytorch>=0.3.0

# Data Processing
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
scikit-image>=0.18.0

# Data Augmentation
albumentations>=1.3.0

# Machine Learning Utilities
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Bars
tqdm>=4.62.0

# Configuration and Utilities
pyyaml>=6.0
dataclasses-json>=0.5.0

# Jupyter (for development)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional: For ONNX export
# onnx>=1.12.0
# onnxruntime>=1.12.0

# Optional: For mixed precision training
# apex (install from GitHub)

# Optional: For advanced optimizers
# timm>=0.6.0

# Development Tools (optional)
# black>=22.0.0
# flake8>=4.0.0
# pytest>=7.0.0
